add canadian univeress to urban centers
